big update

+++ Libs/assimp/include/assimp/types.h
@@ -358,11 +358,11 @@ struct aiString {
         if (!len) {
             return;
         }
-        if (length + len >= MAXLEN) {
+        if (length + (size_t)len >= MAXLEN) {
             return;
         }
 
-        memcpy(&data[length], app, len + 1);
+        memcpy(&data[length], app, (size_t)len + 1);
         length += len;
     }
